service/requirements.txt
click==8.1.6
Flask==2.3.2
Flask_JWT_Extended==4.5.2
pandas==2.2.2
psycopg2_binary==2.9.6
Werkzeug==2.3.6
